# contracts/control-api.openapi.yaml
openapi: 3.1.0
...
info:
  title: AstrLink Local Control API
  version: 1.0.0
  summary: Stable local contract between the AstrLink desktop application and Core.
  description: |
    This API is a local control plane; it is not part of the public inference API.
    Core binds the inference plane to `127.0.0.1:8317` by default and binds this
    control plane to the ephemeral address `127.0.0.1:0`. On startup Core writes
    exactly one ready JSON line to stdout containing the selected `control_url`.
    Launchers must use that value rather than guessing a port.

    The bootstrap health, version, and capability operations are intentionally
    unauthenticated and disclose no configuration. Every operation that reads or
    mutates user configuration, events, or request records requires the per-start
    local control token. The token must be delivered through an inherited pipe or
    equivalent protected IPC; it must not be placed in process arguments or the
    environment.

    Request and response bodies are excluded from ordinary request records. The two
    audit-body switches are independent, default to false, and can only be enabled
    through this authenticated control API after explicit risk acknowledgement.

    The three bootstrap GET operations are always available. Authenticated Service
    CRUD, service-scoped subscription OAuth, priority Route CRUD, persistent local
    access-token operations, the fixed
...
    capability claim.
servers:
...
    variables:
...
security:
  - controlToken: []
...
paths:
...
  /control/v1/services:
    x-astrlink-implementation-status: implemented
    get:
      tags: [Services]
      operationId: listServices
      summary: List every configured API service, including subscription and HTTP variants.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: kind
          in: query
          schema:
            $ref: "#/components/schemas/ServiceKind"
      responses:
        "200":
          description: Service page. Credential secrets and OAuth tokens are never returned.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServicePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Services]
      operationId: createService
      summary: Create one independently addressable API service.
      description: |
        `codex_subscription` creates a disconnected service that is authorized
        through its service-scoped OAuth operation. HTTP kinds such as `newapi`
        carry an `http` connection and optional write-only credential.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceCreate"
      responses:
        "201":
          description: Service created.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            ETag:
              $ref: "#/components/headers/ETag"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /control/v1/service-model-probes:
    x-astrlink-implementation-status: implemented
    post:
      tags: [Services]
      operationId: probeDraftServiceModels
      summary: Discover models using an unsaved HTTP service draft.
      description: A service_id allows an omitted draft credential to reuse that service's stored credential, even when the draft changes the HTTP URL or authentication scheme. An explicit draft credential takes precedence.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftServiceModelProbe"
      responses:
        "200":
          description: Upstream model identifiers without credential material.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceModelProbe"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
...
  /control/v1/access-tokens:
    x-astrlink-implementation-status: implemented
    get:
      tags: [Access Tokens]
      operationId: listAccessTokens
      summary: List persistent local inference access-token metadata.
      responses:
        "200":
          description: Access-token metadata. Secret values and hashes are never included.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Access Tokens]
      operationId: createAccessToken
      summary: Create a persistent local inference access token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessTokenCreate"
      responses:
        "201":
          description: Token created. The complete secret is returned only by this explicit operation and the reveal operation.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenSecret"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /control/v1/routes:
    x-astrlink-implementation-status: implemented
    get:
      tags: [Routes]
      operationId: listRoutes
      summary: List routes in deterministic evaluation order.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: enabled
          in: query
          schema:
            type: boolean
      deprecated: true
      responses:
        '410':
          description: Legacy routing feature retired; use service order and routing settings
        "200":
          description: Route page.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Routes]
      operationId: createRoute
      summary: Create a route.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteCreate"
      deprecated: true
      responses:
        '410':
          description: Legacy routing feature retired; use service order and routing settings
        "201":
          description: Route created.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            ETag:
              $ref: "#/components/headers/ETag"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /control/v1/policies:
    get:
      x-astrlink-implementation-status: implemented
      tags: [Policies]
      operationId: listPolicies
      summary: List the fixed MVP privacy policy.
      responses:
        "200":
          description: Policy page.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      x-astrlink-implementation-status: planned
      tags: [Policies]
      operationId: createPolicy
      summary: Create a policy.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyCreate"
      responses:
        "201":
          description: Policy created.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            ETag:
              $ref: "#/components/headers/ETag"
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /control/v1/privacy/regex-builtin-rules:
    get:
      x-astrlink-implementation-status: implemented
      tags: [Policies]
      operationId: listPrivacyRegexBuiltinRules
      summary: Read the built-in Regex detector catalog.
      description: |
        Returns the fixed kind and pattern pairs used by the builtin Regex
        detector. Intended to seed custom_regex_rules; the catalog itself is
        not editable. Custom mode compiles patterns only and does not reuse
        built-in validators such as Luhn or IBAN checks.
      responses:
        "200":
          description: Built-in Regex rules.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyRegexBuiltinRulesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /control/v1/privacy-model-catalog:
    get:
      x-astrlink-implementation-status: implemented
      tags: [Privacy Model]
      operationId: listPrivacyModelCatalog
      summary: List versioned built-in model choices and their CPU variants.
      responses:
        "200":
          description: Built-in catalog metadata. Model weights are not bundled.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyModelCatalog"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /control/v1/privacy-models/probe:
    post:
      x-astrlink-implementation-status: implemented
      tags: [Privacy Model]
      operationId: probePrivacyModel
      summary: Inspect a public Hugging Face repository without loading model weights.
      description: Content-derived identities matching `local/model-<12 lowercase hex>` are reserved for the mounted-directory probe and return a conflict here until that directory is probed again. Other repositories owned by `local` remain valid public repositories.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrivacyModelProbeRequest"
      responses:
        "200":
          description: Immutable revision, detected ONNX variants, and label-mapping requirements.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyModelProbe"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
  /control/v1/privacy-models/local/probe:
    post:
      x-astrlink-implementation-status: implemented
      tags: [Privacy Model]
      operationId: probeLocalPrivacyModel
      summary: Inspect a model package from an already-mounted local path.
      description: The path may identify a package directory or one ONNX file within it. A file probe pins only that model variant plus its required config, tokenizer, and external data. The source path is used only for the authenticated probe and is never returned, logged, or persisted. Installation copies pinned assets into AstrLink private storage before execution.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrivacyModelLocalProbeRequest"
      responses:
        "200":
          description: Content-pinned local package identity, detected ONNX variants, and label-mapping requirements.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyModelProbe"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /control/v1/privacy-models:
    get:
      x-astrlink-implementation-status: implemented
      tags: [Privacy Model]
      operationId: listPrivacyModelInstallations
      summary: List persistent model-and-variant installations and current progress.
      responses:
        "200":
          description: Installed, downloading, or failed model variants.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyModelInstallationList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      x-astrlink-implementation-status: implemented
      tags: [Privacy Model]
      operationId: installPrivacyModel
      summary: Start or retry one immutable model-and-variant installation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrivacyModelInstallRequest"
      responses:
        "202":
          description: Download accepted.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrivacyModelInstallation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "502":
          $ref: "#/components/responses/BadGateway"
...
components:
  securitySchemes:
    controlToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-local-control-token
      description: Per-start local token delivered over protected IPC; never persist it in configuration or logs.
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor returned by the previous page. It must not be parsed by clients.
      schema:
        type: string
        minLength: 1
        maxLength: 512
...
  headers:
    NoStore:
      description: Sensitive local-control responses must not be cached.
      schema:
        type: string
        examples: [no-store]
    Location:
      description: Relative path of the created resource.
      schema:
        type: string
        format: uri-reference
    ETag:
      description: Strong entity tag for the current resource representation.
      schema:
        type: string
  responses:
    BadRequest:
      description: Malformed request, query, or cursor.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid local control token.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: The operation conflicts with current state or references.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    UnprocessableEntity:
      description: The request is well-formed but violates a contract or safety invariant.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: Local privacy policy or safety engine is temporarily unavailable.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    BadGateway:
      description: Public model metadata could not be fetched or verified.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    GatewayTimeout:
      description: The upstream model-discovery request exceeded its deadline.
      headers:
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ServiceKind:
      type: string
      enum:
        - codex_subscription
        - claude_subscription
        - opencode_go
        - opencode_zen
        - kimi_coding
        - glm_coding
        - minimax_coding
        - newapi
        - openai
        - anthropic
        - gemini
        - openai_compatible
        - custom
      description: >
        Product-level service type. Codex and Claude subscriptions use OAuth.
        Coding plans and other HTTP services use provider-issued API keys.
...
    Service:
      type: object
      required: [id, name, kind, enabled, models, capabilities, created_at, updated_at]
      properties:
        failure_policy:
          $ref: "#/components/schemas/FailurePolicy"
        id:
          $ref: ./protocol-capabilities.schema.json#/$defs/ResourceId
        name:
          type: string
          minLength: 1
          maxLength: 128
        kind:
          $ref: "#/components/schemas/ServiceKind"
        enabled:
          type: boolean
        models:
          $ref: "#/components/schemas/ServiceModels"
        capabilities:
          type: array
          items:
            $ref: ./protocol-capabilities.schema.json#/$defs/ServiceCapability
        http:
          $ref: "#/components/schemas/HTTPServiceConnection"
        subscription:
          $ref: "#/components/schemas/SubscriptionServiceConnection"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
      oneOf:
        - properties:
            kind:
              enum: [codex_subscription, claude_subscription]
          required: [subscription]
          not:
            required: [http]
        - properties:
            kind:
              $ref: ./protocol-capabilities.schema.json#/$defs/HTTPServiceKind
          required: [http]
          not:
            required: [subscription]
...
    ServiceCreate:
      type: object
      required: [name, kind]
      properties:
        failure_policy:
          $ref: "#/components/schemas/FailurePolicy"
        name:
          type: string
          minLength: 1
          maxLength: 128
        kind:
          $ref: "#/components/schemas/ServiceKind"
        enabled:
          type: boolean
          default: true
        models:
          $ref: "#/components/schemas/ServiceModels"
        http:
          $ref: "#/components/schemas/HTTPServiceConnectionInput"
        capabilities:
          type: array
          items:
            $ref: ./protocol-capabilities.schema.json#/$defs/ServiceCapability
      additionalProperties: false
      oneOf:
        - properties:
            kind:
              enum: [codex_subscription, claude_subscription]
          not:
            anyOf:
              - required: [http]
              - required: [capabilities]
        - properties:
            kind:
              $ref: ./protocol-capabilities.schema.json#/$defs/HTTPServiceKind
          required: [http, capabilities]
...
    DraftServiceModelProbe:
      type: object
      required: [kind, http, protocol]
      properties:
        service_id:
          $ref: ./protocol-capabilities.schema.json#/$defs/ResourceId
        kind:
          $ref: ./protocol-capabilities.schema.json#/$defs/HTTPServiceKind
        http:
          $ref: "#/components/schemas/HTTPServiceConnectionInput"
        protocol:
          type: string
          enum: [openai.models, google.models]
      additionalProperties: false
...
    ServicePage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        next_cursor:
          type: [string, "null"]
          maxLength: 512
      additionalProperties: false
...
    ServiceModelProbe:
      type: object
      required: [protocol, model_ids]
      properties:
        service_id:
          $ref: ./protocol-capabilities.schema.json#/$defs/ResourceId
        protocol:
          type: string
          enum: [openai.models, google.models]
        model_ids:
          type: array
          maxItems: 10000
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 256
      additionalProperties: false
...